+++ core/src/include/FieldAdvection.hpp
@@ -28,18 +28,18 @@ public:
      * @returns A reference to the updated ModelArray.
      */
     static ModelArray& advectField(ModelArray& field, const TimestepTime& tst,
-        double lowerLimit = -std::numeric_limits<double>::infinity(),
-        double upperLimit = std::numeric_limits<double>::infinity());
+        FloatType lowerLimit = -std::numeric_limits<FloatType>::infinity(),
+        FloatType upperLimit = std::numeric_limits<FloatType>::infinity());
 
 #ifdef USE_KOKKOS
     static void advectField(const DeviceViewMA& field, const TimestepTime& tst,
-        double lowerLimit = -std::numeric_limits<double>::infinity(),
-        double upperLimit = std::numeric_limits<double>::infinity());
+        FloatType lowerLimit = -std::numeric_limits<FloatType>::infinity(),
+        FloatType upperLimit = std::numeric_limits<FloatType>::infinity());
 
 /*    template<typename ExecSpace>
     static void advectField(ExecSpace execSpace, const DeviceViewMA& field, const TimestepTime& tst,
-        double lowerLimit = -std::numeric_limits<double>::infinity(),
-        double upperLimit = std::numeric_limits<double>::infinity())
+        FloatType lowerLimit = -std::numeric_limits<FloatType>::infinity(),
+        FloatType upperLimit = std::numeric_limits<FloatType>::infinity())
     {}*/
 #endif
 };
