+++ core/test/MonthlyCubicBSpline_test.cpp
@@ -19,11 +19,11 @@ TEST_SUITE_BEGIN("MonthlyCubicBSpline");
  * leap year). */
 TEST_CASE("Test cyclicity")
 {
-    const std::vector<double> testTable = { 1., 2., 3., 4., 5., 6., 7., 8., 9., 10., 11., 12. };
+    const std::vector<FloatType> testTable = { 1., 2., 3., 4., 5., 6., 7., 8., 9., 10., 11., 12. };
 
     monthlyCubicBSpline splineTest(testTable);
 
-    const double prec = 1e-6;
+    const FloatType prec = 1e-6;
     REQUIRE(splineTest(0, 0) == doctest::Approx(splineTest(365., 0)).epsilon(prec));
     REQUIRE(splineTest(0, 1) == doctest::Approx(splineTest(366., 1)).epsilon(prec));
 }
@@ -35,17 +35,17 @@ TEST_CASE("Test cyclicity")
 TEST_CASE("Albedo values")
 {
     // Monthly snow albedo from Maykut and Untersteiner (1971)
-    const std::vector<double> albedoTable
+    const std::vector<FloatType> albedoTable
         = { 0.85, 0.85, 0.83, 0.81, 0.82, 0.78, 0.64, 0.69, 0.84, 0.85, 0.85, 0.85 };
     //      Jan,  Feb,  Mar,  Apr,  Mai,  Jun,  Jul,  Aug,  Sept, Oct,  Nov,  Dec
 
     monthlyCubicBSpline splineTest(albedoTable);
 
-    const double prec = 5e-4;
+    const FloatType prec = 5e-4;
 
     // Loop over the year by mid-month days-of-year (which can be a fraction of a day).
     // Start on January 14th at 5:14:33 and assume each month is 1/12th of a Gregorian year.
-    double dayOfYear = 365.2425 / 24;
+    FloatType dayOfYear = 365.2425 / 24;
     for (auto& element : albedoTable) {
         REQUIRE(splineTest(dayOfYear, 0) == doctest::Approx(element).epsilon(prec));
         dayOfYear += 365.2425 / 12;
