+++ physics/src/BenchmarkCoordinates.cpp
@@ -7,8 +7,8 @@
 namespace Nextsim {
 
 bool BenchmarkCoordinates::isInitialized = false;
-double BenchmarkCoordinates::dx = 25000.;
-double BenchmarkCoordinates::dy = 25000.;
+FloatType BenchmarkCoordinates::dx = 25000.;
+FloatType BenchmarkCoordinates::dy = 25000.;
 size_t BenchmarkCoordinates::m_nx;
 size_t BenchmarkCoordinates::m_ny;
 HField BenchmarkCoordinates::m_x(ModelArray::Type::H);
@@ -27,9 +27,9 @@ void BenchmarkCoordinates::setData()
         dy = 512e3 / m_ny;
 
         for (size_t j = 0; j < m_ny; ++j) {
-            double yVal = j * dy;
+            FloatType yVal = j * dy;
             for (size_t i = 0; i < m_nx; ++i) {
-                double xVal = i * dx;
+                FloatType xVal = i * dx;
                 m_x(i, j) = xVal;
                 m_y(i, j) = yVal;
             }
